Bandit is stuck on ADC setup again... shouldn't be

+++ src/Bandit_Template/Bandit_Debug.h
@@ -15,7 +15,7 @@
 struct BANDIT_CORE_1_DEBUG_REPORT {
     uint32_t    epoch_core_1_boot_us;           // epoch us elapsed until Core 1 boots
     uint16_t    adc_init_cfr_readback;          // CFR readback value, should be 2624
-    uint16_t    adc_init_attempts;              // Number of attempts at successful configuration of ADC
+    uint32_t    adc_init_attempts;              // Number of attempts at successful configuration of ADC
     uint32_t    epoch_ads_init_complete_us;     // epoch us elapsed till ADC configuration complete
     uint8_t     dac_cal_state;                  // See enum 1
     uint16_t    dac_cal_value;                  // Value written to ADC (without << 3 shift)
